updates: - organize file - move history to modal details - make shipping modal
