+++ FrontendRSReact/src/components/orderPages/ModalOrderDetails.js
@@ -1,11 +1,19 @@
 import React, { useState, useRef } from "react";
-import Modal from "../Modal";
-import { FaTrash, FaInfoCircle, FaHandshake, FaCheck } from "react-icons/fa";
-import { useNavigate } from "react-router-dom";
+import {
+  MdInventory,
+  MdMenu,
+  MdInfo,
+  MdHandshake,
+  MdDelete,
+  MdSettingsBackupRestore,
+  MdCheck,
+  MdLocalShipping,
+} from "react-icons/md";
 import { useSelector } from "react-redux";
-import axios, { all } from "axios";
+import axios from "axios";
 import printOrderItem from "./printOrderItem";
 import { useEffect } from "react";
+import PrintBeritaAcara from "./PrintBeritaAcara";
 
 const ModalOrderDetails = ({
   onClose,
@@ -13,6 +21,10 @@ const ModalOrderDetails = ({
   handleHistory,
   handleDetailProduct,
   handleOffer,
+  handleOfferAccepted,
+  handleRefund,
+  handleShipping,
+  handleConfirm,
   handleOpenSubmitModal,
   handlePayoutDetail,
   setShowActionToast,
@@ -20,7 +32,6 @@ const ModalOrderDetails = ({
   setActionToastBody,
   fetchData,
 }) => {
-  const navigate = useNavigate();
   const role = useSelector((state) => state.auth.role);
   const id = useSelector((state) => state.auth.id);
   const [negotiable, setNegotiable] = useState(true);
@@ -30,20 +41,35 @@ const ModalOrderDetails = ({
     (orderItem) => orderItem.status === "ACCEPTED"
   );
 
+  const allItemsChecked = selectedOrder.orderItems.every(
+    (orderItem) => orderItem.status === "CHECKED"
+  );
+
+  useEffect(() => {
+    console.log("modal order details", window.detailsModal.open);
+    if (!window.detailsModal.open) {
+      window.detailsModal.showModal();
+    }
+  }, [selectedOrder]);
+
   useEffect(() => {
     if (
       selectedOrder.orderItems.every((order) => order.status === "OFFER") ||
-      selectedOrder.orderItems.every((order) => order.status === "ACCEPTED")
+      selectedOrder.orderItems.every((order) => order.status === "ACCEPTED") ||
+      selectedOrder.orderItems.every((order) => order.status === "REFUND") ||
+      selectedOrder.orderItems.every((order) => order.status === "CHECKED")
     ) {
       setNegotiable(false);
+    }
+    if (
+      selectedOrder.status === "ORDER" ||
+      selectedOrder.status === "NEGOTIATION"
+    ) {
+      setNegotiable(true);
     }
   }, [selectedOrder]);
 
   console.log("selected order", selectedOrder.orderItems);
-
-  const handleQuantityChange = (orderItemId, newQuantity) => {
-    console.log("quantity changed", orderItemId, newQuantity);
-  };
 
   const handleDeleteOrderItem = (orderItemId) => {
     const confirmed = window.confirm(
@@ -149,6 +175,47 @@ const ModalOrderDetails = ({
     onClose();
     window.location.reload();
   };
+
+  const handlePrintBeritaAcara = () => {
+    axios
+      .put(
+        `http://rsudsamrat.site:8080/pengadaan/dev/v1/orders/${selectedOrder.id}/status`,
+        {
+          status: "PAYMENT",
+        }
+      )
+      .then((response) => {
+        axios
+          .get(`http://rsudsamrat.site:8080/employee`)
+          .then((response) => {
+            // map the response.data and get the id of each employee and push it to employeeList
+            let employeeList = [];
+            let employeeNameList = [];
+            response.data.map((employee) => {
+              employeeList.push(employee.id);
+              employeeNameList.push(employee.name);
+            });
+            console.log(employeeList);
+            axios
+              .post(`http://rsudsamrat.site:8990/api/v1/notifikasi`, {
+                sender: role,
+                senderId: id,
+                receiver: employeeNameList,
+                receiverId: employeeList,
+                message: `THIS ORDER IS PAYMENT, BY ${role}`,
+              })
+              .then((response) => {
+                console.log("berhasil mengirim notifikasi ke KEUANGAN");
+              })
+              .catch((err) => console.log(err));
+          })
+          .catch((err) => console.log(err));
+        console.log(response);
+      })
+      .catch((err) => console.log(err));
+    onClose();
+  };
+
   const fileInputRef = useRef(null);
 
   const handleUploadClick = () => {
@@ -194,126 +261,175 @@ const ModalOrderDetails = ({
   //   }
   // }, [selectedOrder.Order])
 
-  return (
-    <Modal title="SubOrder Details" onClose={onClose} modalSize="xl">
-      <div className="modal-body">
-        <h4>Order Items:</h4>
-        <table className="table">
-          <thead>
-            <tr>
-              <th>OrderId</th>
-              <th>Product UUID</th>
-              <th>Vendor Name</th>
-              <th>Product Name</th>
-              <th>Product Price</th>
-              <th>bidprice</th>
-              <th>STATUS</th>
-              <th>Quantity</th>
-              <th>Action</th>
-            </tr>
-          </thead>
-          <tbody>
-            {(() => {
-              const vendorItemsMap = new Map();
+  const ProductItem = ({ orderItem }) => {
+    console.log("--productItem--", orderItem);
 
-              selectedOrder.orderItems.forEach((orderItem) => {
-                const vendorName = orderItem.product.vendor.name;
+    const calculateTotalPrice = () => {
+      // get total price by multiplying quantity with price
+      const totalPrice = orderItem.quantity * orderItem.product.price;
+      return totalPrice;
+    };
 
-                if (vendorItemsMap.has(vendorName)) {
-                  vendorItemsMap.get(vendorName).push(orderItem);
-                } else {
-                  vendorItemsMap.set(vendorName, [orderItem]);
-                }
-              });
+    return (
+      <tr>
+        <td className="font-bold">{orderItem.id}</td>
+        <td className="font-medium text-primary-1">
+          {orderItem.product.name}{" "}
+          <span className="text-sm text-black">x{orderItem.quantity}</span>
+        </td>
+        <td>
+          Rp{" "}
+          <span>
+            {orderItem.totalAmount === 0
+              ? calculateTotalPrice()
+              : orderItem.totalAmount}
+          </span>
+        </td>
+        <td className="cursor-pointer hover:text-primary-1 transition-all duration-300 ease-in-out font-semibold">
+          {orderItem.status}
+        </td>
+        <td>
+          <div className="dropdown dropdown-bottom dropdown-end">
+            <label
+              tabIndex={0}
+              className="text-xl text-white btn border-primary-1 bg-primary-1 hover:bg-primary-2 hover:border-primary-2"
+            >
+              <MdMenu />
+            </label>
+            <ul
+              tabIndex={0}
+              className="dropdown-content z-[1] menu p-2 shadow bg-base-100 rounded-box w-52"
+            >
+              {selectedOrder.status === "SHIPPING" && (
+                <li>
+                  <a onClick={() => handleShipping(orderItem.id)}>
+                    <MdLocalShipping className="text-xl text-orange-500" />
+                    Shipping
+                  </a>
+                </li>
+              )}
+              <hr />
+              {selectedOrder.status === "ORDER" ||
+                (selectedOrder.status === "NEGOTIATION" && (
+                  <li>
+                    <a onClick={() => handleOfferAccepted(orderItem.id)}>
+                      <MdCheck className="text-xl text-success" />
+                      Accept Order
+                    </a>
+                  </li>
+                ))}
+              {(orderItem.status === "ACCEPTED" ||
+                orderItem.status === "RESEND") &&
+                (selectedOrder.status === "SHIPPING" ||
+                  selectedOrder.status === "CHECKING") && (
+                  <>
+                    <li>
+                      <a onClick={() => handleConfirm(orderItem.id)}>
+                        <MdCheck className="text-xl text-success" />
+                        Confirm
+                      </a>
+                    </li>
+                    <hr />
+                    <li>
+                      <a onClick={() => handleRefund(orderItem.id)}>
+                        <MdSettingsBackupRestore className="text-xl text-slate-500" />
+                        Refund
+                      </a>
+                    </li>
+                  </>
+                )}
+              <hr />
+              <li>
+                <a
+                  onClick={() =>
+                    handleDetailProduct(orderItem.product.productuuid)
+                  }
+                >
+                  <MdInfo className="text-xl text-slate-500" />
+                  Info
+                </a>
+              </li>
+              <hr />
+              {negotiable && (
+                <li>
+                  <a onClick={() => handleOffer(orderItem.id)}>
+                    <MdHandshake className="text-xl text-success" />
+                    Negotiation
+                  </a>
+                </li>
+              )}
+              <hr />
+              <li>
+                <a onClick={() => handleDeleteOrderItem(orderItem.id)}>
+                  <MdDelete className="text-xl text-red-500" />
+                  Delete
+                </a>
+              </li>
+            </ul>
+          </div>
+        </td>
+      </tr>
+    );
+  };
 
-              return [...vendorItemsMap].map(([vendorName, orderItems]) => (
-                <React.Fragment key={vendorName}>
-                  <tr>
-                    <td colSpan="9">
-                      <strong>{vendorName}</strong>
-                    </td>
-                  </tr>
-                  {orderItems.map((orderItem) => (
-                    <tr key={orderItem.id}>
-                      <td>{orderItem.id}</td>
-                      <td>{orderItem.product.productuuid}</td>
-                      <td>{vendorName}</td>
-                      <td>{orderItem.product.name}</td>
-                      <td>{orderItem.product.price * orderItem.quantity}</td>
-                      <td>{orderItem.bidPrice}</td>
-                      <td
-                        onClick={() => handleHistory(orderItem.id)}
-                        className="history-click"
-                      >
-                        {orderItem.status}
-                      </td>
-                      <td>{orderItem.quantity}</td>
-                      <td>
-                        <button
-                          className="btn btn-sm btn-clear text-danger"
-                          onClick={() => handleDeleteOrderItem(orderItem.id)}
-                        >
-                          <FaTrash />
-                        </button>
-                        <button
-                          className="btn btn-sm btn-clear text-info"
-                          onClick={() =>
-                            handleDetailProduct(orderItem.product.productuuid)
-                          }
-                        >
-                          <FaInfoCircle />
-                        </button>
-                        {negotiable && (
-                          <button
-                            className="btn btn-sm btn-clear text-success"
-                            onClick={() => handleOffer(orderItem.id)}
-                          >
-                            <FaHandshake />
-                          </button>
-                        )}
-                        {/* {orderItem.status === "ACCEPTED" && (
-                          <button
-                            className="btn btn-sm btn-clear text-success"
-                            onClick={() => handleOpenSubmitModal(orderItem)}
-                          >
-                            <FaCheck />
-                          </button>
-                        )} */}
-                      </td>
-                    </tr>
-                  ))}
-                  {orderItems.every(
-                    (orderItem) => orderItem.status === "ACCEPTED"
-                  ) && (
-                    <tr>
-                      <td colSpan="9">
-                        <button
-                          className="btn btn-sm btn-success"
-                          onClick={() =>
-                            handleSetStatusValidating("VALIDATING")
-                          }
-                        >
-                          Change Status
-                        </button>
-                      </td>
-                    </tr>
-                  )}
-                </React.Fragment>
-              ));
-            })()}
-          </tbody>
-        </table>
-        <div className="d-flex gap-2">
-          {role === "PP" && (
+  const productPerVendorItem = () => {
+    const vendorItemsMap = new Map();
+
+    selectedOrder.orderItems.forEach((orderItem) => {
+      const vendorName = orderItem.product.vendor.name;
+
+      if (vendorItemsMap.has(vendorName)) {
+        vendorItemsMap.get(vendorName).push(orderItem);
+      } else {
+        vendorItemsMap.set(vendorName, [orderItem]);
+      }
+    });
+
+    return [...vendorItemsMap].map(([vendorName, orderItems], index) => (
+      <React.Fragment key={index}>
+        <div className="flex flex-col gap-2 ">
+          <h2 className="font-medium text-xl text-slate-600">Best Vendor</h2>
+          <hr />
+          <table className="table table-pin-rows">
+            {/* head */}
+            <thead>
+              <tr>
+                <th>Order ID</th>
+                <th>Product Name</th>
+                <th>Final Price</th>
+                <th>Status</th>
+                <th>Actions</th>
+              </tr>
+            </thead>
+            <tbody>
+              {/* body */}
+              {orderItems.map((orderItem, index) => (
+                <ProductItem orderItem={orderItem} key={index} />
+              ))}
+              {/* {productItem()} */}
+            </tbody>
+          </table>
+        </div>
+        <div className="flex gap-2 my-2">
+          {orderItems.every((orderItem) => orderItem.status === "ACCEPTED") &&
+            selectedOrder.status === "NEGOTIATION" && (
+              <button
+                className="text-white btn btn-sm border-primary-1 bg-primary-1 hover:bg-primary-2 hover:border-primary-2"
+                onClick={() => handleSetStatusValidating("VALIDATING")}
+              >
+                Change Status
+              </button>
+            )}
+          {role === "PPKOM" && selectedOrder.status === "VALIDATING" && (
             <>
               <button
-                className="btn btn-secondary"
+                className="text-white btn btn-sm border-primary-1 bg-primary-1 hover:bg-primary-2 hover:border-primary-2"
                 onClick={() => handleSetStatusNego("NEGOTIATION")}
               >
                 Cancel Negotiation
               </button>
               <button
-                className="btn btn-danger"
+                className="text-white btn btn-sm border-primary-1 bg-primary-1 hover:bg-primary-2 hover:border-primary-2"
                 onClick={() => handleSetStatusCancel("CANCEL")}
               >
                 Cancel Order
@@ -321,64 +437,106 @@ const ModalOrderDetails = ({
             </>
           )}
         </div>
-      </div>
-      <div className="modal-footer">
-        {/* Remove Comment to only enable Print button on 'PP' Role */}
-        {/* {role === "PP" ? ( */}
-        {allItemsAccepted && (
-          <button
-            type="button"
-            className="btn btn-secondary"
-            onClick={() => printOrderItem(selectedOrder.orderItems)}
-          >
-            Print
-          </button>
-        )}
-        {allItemsAccepted && (
-          <div>
-            <input
-              type="file"
-              style={{ display: "none" }}
-              ref={fileInputRef}
-              onChange={handleFileChange}
-            />
-            <button
-              onClick={handleUploadClick}
-              className="btn btn-secondary"
-              disabled={selectedFiles.length === 1}
-            >
-              Upload Surat
+        <hr />
+      </React.Fragment>
+    ));
+  };
+
+  return (
+    <dialog id="detailsModal" className="modal">
+      <div className="modal-box max-w-4xl">
+        <form method="dialog">
+          {/* Header */}
+          <div className="flex items-center justify-between mb-3">
+            <div className="flex items-center gap-2">
+              <MdInventory className="text-2xl text-primary-1" />
+              <h3 className="text-xl font-bold">Order Details</h3>
+            </div>
+            <button className="btn btn-sm btn-circle btn-ghost absolute right-2 top-2">
+              ✕
             </button>
           </div>
-        )}
-        {/* ) : (
-          <> */}
-        <button
-          type="button"
-          className="btn btn-secondary"
-          onClick={handlePayoutDetail}
-        >
-          Payout Detail
-        </button>
-        <button type="button" className="btn btn-secondary">
-          Check Status
-        </button>
-        {/* </>
-        )} */}
-      </div>
-      {selectedFiles && (
-        <div>
-          {selectedFiles.map((file, index) => (
-            <div style={{ display: "flex" }}>
-              <p key={index}>{file.name}</p>
-              <button className="btn btn-secondary" onClick={handleUploadFile}>
-                Upload
+
+          {/* Item list/vendor Item */}
+          {productPerVendorItem()}
+        </form>
+
+        {/* Footer */}
+        <div className="modal-action mt-12">
+          {selectedFiles && (
+            <div>
+              {selectedFiles.map((file, index) => (
+                <div style={{ display: "flex" }} key={index}>
+                  <p key={index}>{file.name}</p>
+                  <button
+                    className="text-white btn border-primary-1 bg-primary-1 hover:bg-primary-2 hover:border-primary-2"
+                    onClick={handleUploadFile}
+                  >
+                    Upload
+                  </button>
+                </div>
+              ))}
+            </div>
+          )}
+          {allItemsChecked && (
+            <div>
+              <input
+                type="file"
+                style={{ display: "none" }}
+                ref={fileInputRef}
+                onChange={handleFileChange}
+              />
+              <button
+                onClick={() => {
+                  handleUploadClick();
+                  handlePrintBeritaAcara();
+                }}
+                className="text-white btn border-primary-1 bg-primary-1 hover:bg-primary-2 hover:border-primary-2"
+                disabled={selectedFiles.length === 1}
+              >
+                Upload Surat
               </button>
             </div>
-          ))}
+          )}
+          <button
+            onClick={handleHistory}
+            className="text-white btn border-primary-1 bg-primary-1 hover:bg-primary-2 hover:border-primary-2"
+          >
+            View History
+          </button>
+          <button
+            onClick={handlePayoutDetail}
+            className="text-white btn border-primary-1 bg-primary-1 hover:bg-primary-2 hover:border-primary-2"
+          >
+            Payout Details
+          </button>
+          <button className="text-white btn border-primary-1 bg-primary-1 hover:bg-primary-2 hover:border-primary-2">
+            Check Status
+          </button>
+          {allItemsAccepted && (
+            <button
+              type="button"
+              className="text-white btn border-primary-1 bg-primary-1 hover:bg-primary-2 hover:border-primary-2"
+              onClick={() => printOrderItem(selectedOrder.orderItems)}
+            >
+              Print
+            </button>
+          )}
+          {/* display if all the items status is CHECKED */}
+          {allItemsChecked && (
+            <button
+              type="button"
+              className="text-white btn border-primary-1 bg-primary-1 hover:bg-primary-2 hover:border-primary-2"
+              onClick={() => {
+                PrintBeritaAcara(selectedOrder);
+              }}
+            >
+              Print Berita Acara
+            </button>
+          )}
         </div>
-      )}
-    </Modal>
+      </div>
+    </dialog>
   );
 };
