+++ FrontendRSReact/src/config/index.js
@@ -12,7 +12,6 @@ import Footer from "../components/Footer";
 // import "bootstrap/dist/css/bootstrap.min.css";
 import Vendorpages from "../pages/Vendorpages";
 import Orderpages from "../pages/Orderpages";
-import OrderpagesNew from "../pages/OrderpagesNew";
 import Notificationpages from "../pages/Notificationpages";
 import Profilpages from "../pages/Profilpages";
 import SignInpages from "../pages/SignInpages";
@@ -20,7 +19,6 @@ import Vendors from "../pages/Vendor";
 import CommingSoon from "../pages/CommingSoon";
 import Payment from "../pages/Payment";
 import { useSelector } from "react-redux";
-import ProductPagesNew from "../pages/ProductPagesNew";
 import PageContainer from "../components/PageContainer";
 
 const Routers = () => {
@@ -42,10 +40,8 @@ const Routers = () => {
             <>
               <Route path="/" element={<Dashboard />} />
               <Route path="/products" element={<Productpages />} />
-              <Route path="/productsNew" element={<ProductPagesNew />} />
               <Route path="/vendor" element={<Vendorpages />} />
               <Route path="/orders" element={<Orderpages />} />
-              <Route path="/ordersNew" element={<OrderpagesNew />} />
               <Route path="/notifications" element={<Notificationpages />} />
               <Route path="/profile" element={<Profilpages />} />
               <Route path="/Vendors" element={<Vendors />} />
