+++ FrontendRSReact/src/pages/Orderpages.js
@@ -3,16 +3,17 @@ import axios from "axios";
 import { useSelector } from "react-redux";
 import { useLocation } from "react-router-dom";
 
-import ModalHistory from "../components/orderPagesNew/ModalHistory";
-import ModalOrderItem from "../components/orderPagesNew/ModalOrderItem";
-// import ModalAcceptedOffer from "../components/orderPagesNew/ModalAcceptedOffer";
-// import ModalSubmittedOffer from "../components/orderPagesNew/ModalSubmittedOffer";
-import ModalOffer from "../components/orderPagesNew/ModalOffer";
-import ModalPayoutDetails from "../components/orderPagesNew/ModalPayoutDetails";
-import ModalProductDetails from "../components/orderPagesNew/ModalProductDetails";
-import ModalOrderDetails from "../components/orderPagesNew/ModalOrderDetails";
-import ModalRefund from "../components/orderPagesNew/ModalRefund";
-import ModalConfirm from "../components/orderPagesNew/ModalConfirm";
+import ModalHistory from "../components/orderPages/ModalHistory";
+import ModalOrderItem from "../components/orderPages/ModalOrderItem";
+// import ModalAcceptedOffer from "../components/orderPages/ModalAcceptedOffer";
+// import ModalSubmittedOffer from "../components/orderPages/ModalSubmittedOffer";
+import ModalOffer from "../components/orderPages/ModalOffer";
+import ModalPayoutDetails from "../components/orderPages/ModalPayoutDetails";
+import ModalProductDetails from "../components/orderPages/ModalProductDetails";
+import ModalOrderDetails from "../components/orderPages/ModalOrderDetails";
+import ModalRefund from "../components/orderPages/ModalRefund";
+import ModalConfirm from "../components/orderPages/ModalConfirm";
+import ModalShipping from "../components/orderPages/ModalShipping";
 
 const Orderpages = () => {
   const [data, setData] = useState([]);
@@ -35,7 +36,9 @@ const Orderpages = () => {
   const [showRefundModal, setShowRefundModal] = useState(false);
   const [showConfirmModal, setShowConfirmModal] = useState(false);
   const [refund, setRefund] = useState(null);
+  const [shipping, setShipping] = useState(null);
   const [confirm, setConfirm] = useState(null);
+  const [showShippingModal, setShowShippingModal] = useState(false);
 
   const role = useSelector((state) => state.auth.role);
   const idUser = useSelector((state) => state.auth.id);
@@ -143,12 +146,12 @@ const Orderpages = () => {
   };
 
   /* History function */
-  const handleHistory = (orderItemId) => {
+  const handleHistory = () => {
     setShowHistoryModal(true);
     console.log("history modal", showHistoryModal);
     axios
       .get(
-        `http://rsudsamrat.site:8090/api/bid-exchange/bid-items/${selectedOrder.id}/${orderItemId}`
+        `http://rsudsamrat.site:8990/api/bid-exchange/history/${selectedOrder.id}`
       )
       .then((res) => {
         console.log("History", res.data);
@@ -506,6 +509,13 @@ const Orderpages = () => {
     }
   };
 
+  const handleShipping = (orderItemId) => {
+    // set shipping modal open
+    setShowShippingModal(true);
+
+    setShipping(orderItemId);
+  };
+
   return (
     <>
       {/* Order Details Modal */}
@@ -515,6 +525,7 @@ const Orderpages = () => {
           selectedOrder={selectedOrder}
           selectedOrderItem={selectedOrderItem}
           handleHistory={handleHistory}
+          handleShipping={handleShipping}
           handleDetailProduct={handleDetailProduct}
           handleOffer={handleOffer}
           handleOfferAccepted={handleAcceptOrder}
@@ -567,11 +578,20 @@ const Orderpages = () => {
           selectedOrderItem={selectedOrderItem}
         />
       )}
+
       {/* History Modal */}
       {showHistoryModal && (
         <ModalHistory
           history={history}
           onClose={() => setShowHistoryModal(null)}
+        />
+      )}
+
+      {/* History Modal */}
+      {showShippingModal && (
+        <ModalShipping
+          shipping={shipping}
+          onClose={() => setShowShippingModal(null)}
         />
       )}
